fix: export public constants and refactor internal constants closer to their usage
