+++ src/icons/ResizingIcon.tsx
@@ -4,10 +4,8 @@ import Box from '@mui/material/Box';
 
 import React, { FC } from 'react';
 
-import {
-  RESIZING_ICON_LEVEL_BACKGROUND_COLOR,
-  RESIZING_ICON_LEVEL_BACKGROUND_COLOR_FOCUS,
-} from '../constants';
+const RESIZING_ICON_LEVEL_BACKGROUND_COLOR = 300;
+const RESIZING_ICON_LEVEL_BACKGROUND_COLOR_FOCUS = 500;
 
 const StyledIcon = styled(UnfoldMoreIcon)(({ theme }: { theme: Theme }) => ({
   backgroundColor: theme.palette.grey[RESIZING_ICON_LEVEL_BACKGROUND_COLOR],
