+++ src/itemLogin/constants.ts
@@ -1,0 +1,1 @@
+export const FORBIDDEN_TEXT = 'You cannot access this item';
