+++ src/items/constants.ts
@@ -1,0 +1,1 @@
+export const ITEM_MAX_HEIGHT = '70vh';
